+++ android-activity/src/game_activity/mod.rs
@@ -8,8 +8,7 @@ use std::ops::Deref;
 use std::os::raw;
 use std::os::unix::prelude::*;
 use std::ptr::NonNull;
-use std::sync::Arc;
-use std::sync::RwLock;
+use std::sync::{Arc, RwLock};
 use std::time::Duration;
 use std::{ptr, thread};
 
@@ -25,7 +24,7 @@ use ndk::configuration::Configuration;
 use ndk::looper::FdEvent;
 use ndk::native_window::NativeWindow;
 
-use crate::{util, AndroidApp, MainEvent, NativeWindowRef, PollEvent, Rect};
+use crate::{util, AndroidApp, ConfigurationRef, MainEvent, NativeWindowRef, PollEvent, Rect};
 
 mod ffi;
 
@@ -121,15 +120,12 @@ impl AndroidApp {
         // Note: we don't use from_ptr since we don't own the android_app.config
         // and need to keep in mind that the Drop handler is going to call
         // AConfiguration_delete()
-        //
-        // Whenever we get a ConfigChanged notification we synchronize this
-        // config state with a deep copy.
         let config = Configuration::clone_from_ptr(NonNull::new_unchecked((*ptr.as_ptr()).config));
 
         Self {
             inner: Arc::new(RwLock::new(AndroidAppInner {
                 native_app: NativeAppGlue { ptr },
-                config: RwLock::new(config),
+                config: ConfigurationRef::new(config),
                 native_window: Default::default(),
             })),
         }
@@ -153,7 +149,7 @@ unsafe impl Sync for NativeAppGlue {}
 #[derive(Debug)]
 pub struct AndroidAppInner {
     native_app: NativeAppGlue,
-    config: RwLock<Configuration>,
+    config: ConfigurationRef,
     native_window: RwLock<Option<NativeWindow>>,
 }
 
@@ -254,7 +250,7 @@ impl AndroidAppInner {
                                         MainEvent::LostFocus
                                     }
                                     ffi::NativeAppGlueAppCmd_APP_CMD_CONFIG_CHANGED => {
-                                        MainEvent::ConfigChanged
+                                        MainEvent::ConfigChanged {}
                                     }
                                     ffi::NativeAppGlueAppCmd_APP_CMD_LOW_MEMORY => {
                                         MainEvent::LowMemory
@@ -282,11 +278,10 @@ impl AndroidAppInner {
                                 trace!("Calling android_app_pre_exec_cmd({cmd_i})");
                                 ffi::android_app_pre_exec_cmd(native_app.as_ptr(), cmd_i);
                                 match cmd {
-                                    MainEvent::ConfigChanged => {
-                                        *self.config.write().unwrap() =
-                                            Configuration::clone_from_ptr(NonNull::new_unchecked(
-                                                (*native_app.as_ptr()).config,
-                                            ));
+                                    MainEvent::ConfigChanged { .. } => {
+                                        self.config.replace(Configuration::clone_from_ptr(
+                                            NonNull::new_unchecked((*native_app.as_ptr()).config),
+                                        ));
                                     }
                                     MainEvent::InitWindow { .. } => {
                                         let win_ptr = (*native_app.as_ptr()).window;
@@ -354,8 +349,8 @@ impl AndroidAppInner {
         }
     }
 
-    pub fn config(&self) -> Configuration {
-        self.config.read().unwrap().clone()
+    pub fn config(&self) -> ConfigurationRef {
+        self.config.clone()
     }
 
     pub fn content_rect(&self) -> Rect {